requirements.txt
absl-py==2.0.0
ansicon==1.89.0
arrow==1.3.0
asgiref==3.7.2
astunparse==1.6.3
blessed==1.20.0
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.7.1
Django==4.2.7
django-allauth==0.58.1
django-allauth-ui==0.1.5
django-picklefield==3.1
django-q==1.3.9
django-widget-tweaks==1.5.0
filelock==3.13.1
flatbuffers==23.5.26
fsspec==2023.10.0
gast==0.5.4
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.10.0
huggingface-hub==0.19.4
idna==3.4
Jinja2==3.1.2
jinxed==1.2.0
keras==2.15.0
libclang==16.0.6
lxml==4.9.3
Markdown==3.5.1
MarkupSafe==2.1.3
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-pptx==0.6.23
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==3.5.3
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.0
six==1.16.0
sqlparse==0.4.4
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tokenizers==0.15.0
torch==2.1.1
torchaudio==2.1.1
torchvision==0.16.1
tqdm==4.66.1
transformers==4.35.2
types-python-dateutil==2.8.19.14
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.12
Werkzeug==3.0.1
wrapt==1.14.1
XlsxWriter==3.1.9

dj-database-url
psycopg2-binary